Worker utils: Clarified API versions of utils
- Worker (non-Conan) remains in utils
- ConanWorker, ConanAPI, monkeypatch moved to v1
