+++ cruiz/workers/api/v1/arbitrary.py
@@ -9,9 +9,9 @@ from __future__ import annotations
 import multiprocessing
 
 from cruiz.interop.commandparameters import CommandParameters
-
-from cruiz.workers.utils import worker
 from cruiz.interop.message import Message, Success
+
+from . import worker
 
 
 def invoke(queue: multiprocessing.Queue[Message], params: CommandParameters) -> None:
