+++ cruiz/workers/api/v1/create.py
@@ -10,9 +10,10 @@ import multiprocessing
 import typing
 
 from cruiz.interop.commandparameters import CommandParameters
-from cruiz.workers.utils import worker
-from cruiz.workers.utils.formatoptions import format_options
 from cruiz.interop.message import Message, Success
+from cruiz.workers.utils.formatoptions import format_options
+
+from . import worker
 
 
 def invoke(queue: multiprocessing.Queue[Message], params: CommandParameters) -> None:
