+++ cruiz/workers/api/v1/exportpackage.py
@@ -10,8 +10,9 @@ import multiprocessing
 import typing
 
 from cruiz.interop.commandparameters import CommandParameters
-from cruiz.workers.utils import worker
 from cruiz.interop.message import Message, Success
+
+from . import worker
 
 
 def invoke(queue: multiprocessing.Queue[Message], params: CommandParameters) -> None:
