+++ cruiz/workers/api/v1/packagerevisions.py
@@ -9,9 +9,9 @@ from __future__ import annotations
 import multiprocessing
 
 from cruiz.interop.packagerevisionsparameters import PackageRevisionsParameters
-
-from cruiz.workers.utils import worker
 from cruiz.interop.message import Message, Success
+
+from . import worker
 
 
 def invoke(
