+++ cruiz/workers/api/v1/remotesearch.py
@@ -10,9 +10,9 @@ import multiprocessing
 import re
 
 from cruiz.interop.searchrecipesparameters import SearchRecipesParameters
-
-from cruiz.workers.utils import worker
 from cruiz.interop.message import Message, Success
+
+from . import worker
 
 
 def invoke(
